pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "baseball-analysis"
version = "0.1.0"
description = "2D baseball pose analysis pipeline with configurable preprocessing experiments."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "pandas>=2.0",
  "opencv-python>=4.8",
  "pyyaml>=6.0",
  "scipy>=1.10",
  "matplotlib>=3.7",
  "mediapipe>=0.10",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "ruff>=0.4",
]
rtmpose = [
  "mmpose>=1.3",
]

[project.scripts]
baseball-pose = "baseball_pose.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"
